Streamlining all polling services
1. RDP1 logs are saved by the server as files, so the rdp1 polling service should use most of the
logic that is inside of the localfile polling service, which has been renamed filepolling. Thus we
only have to extend very few bits and add protocols for new types of resources that are not yet
supported.
2. In the same vein, the file plugin logic should run most of the stuff that is needed for the
server plugin, we just have to add a process basically. This hints we may want to create something
akin to process vs. threaded plugins, for when we have an independent plc or polling service